// client/src/Avatar/type.ts
/* 
TODO: Update because this is not just for 
Avatars but for Characters 
*/
export type AvatarType = {
  allies?: string[];
  enemies?: string[];
  _id: string;
  photoUrl?: string;
  name: string;
  gender?: string;
  eye?: string;
  hair?: string;
  skin?: string;
  love?: string;
  weapon?: string;
  profession?: string;
  position?: string;
  predecessor?: string;
  affiliation?: string;
  first?: string;
};